+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [0.9.4] - 2026-07-08
+
 ### Added
 
 - Add `HTMLContainerBuilder` factory methods to `ExpandableSectionToggle` for custom wrapper elements
@@ -1154,7 +1156,8 @@ Stay tuned for more to come...
 - for dependency upgrades
 -->
 
-[Unreleased]: https://github.com/patternfly-java/patternfly-java/compare/v0.9.3...HEAD
+[Unreleased]: https://github.com/patternfly-java/patternfly-java/compare/v0.9.4...HEAD
+[0.9.4]: https://github.com/patternfly-java/patternfly-java/compare/v0.9.3...v0.9.4
 [0.9.3]: https://github.com/patternfly-java/patternfly-java/compare/v0.9.2...v0.9.3
 [0.9.2]: https://github.com/patternfly-java/patternfly-java/compare/v0.9.1...v0.9.2
 [0.9.1]: https://github.com/patternfly-java/patternfly-java/compare/v0.9.0...v0.9.1